+++ src/typerANDviewer/viewer.cpp
@@ -20,7 +20,7 @@ int main(int argc, char* argv[])
       exit(1);
   }
 
-  int curNum;
+  double curNum;
   int size=0;
   double time=0;
